+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   - Add UART support for Atlas Scientific PT-1000 sensor
   - Update Korean translations
   - Add measurement retries upon CRC fail for AM2315 sensor ([#246](https://github.com/kizniche/mycodo/issues/246))
+  - Add page error handler that provides full traceback when the Web UI crashes
 
 ### Bugfixes